Update README.md
Moves the red wire from the sensor to the cable of the motion part. Drop the blue wire. the Motion sensor then functions on its own and takes over the runout switch function. This is because otherwise there are not enough ports for the endstops. It works very well, the end switch is not missed.